requirements.txt
# Core dependencies
anthropic>=0.75.0  # Claude API
openpyxl>=3.1.0   # Excel 읽기/쓰기
python-dotenv>=1.0.0  # 환경변수 관리
Pillow>=10.0.0  # 이미지 처리

# CLI
click>=8.1.0

# Web interface
streamlit>=1.28.0  # 웹 UI
altair>=5.0.0  # 차트 시각화

# Feedback & RL
pandas>=2.0.0  # 데이터 분석

# Optional: API server
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Development
# pytest>=7.4.0
# black>=23.0.0
# pylint>=3.0.0
